shuffling files around... not sure how I want to organize
