+++ src/containers/org/components/index.js
@@ -4,8 +4,6 @@
 
 export { default as AddRoleToOrgModal } from './AddRoleToOrgModal';
 export { default as CreateOrgModal } from './CreateOrgModal';
-export { default as DataSetPermissionsCard } from './DataSetPermissionsCard';
-export { default as DataSetPermissionsContainer } from './DataSetPermissionsContainer';
 export { default as PermissionsPanel } from './PermissionsPanel';
 export { default as RemoveMemberFromOrgModal } from './RemoveMemberFromOrgModal';
 export { default as RemoveRoleFromMemberModal } from './RemoveRoleFromMemberModal';
